+++ client/src/Product/index.tsx
@@ -5,7 +5,7 @@ import {
   useRouteMatch,
 } from "react-router-dom"
 import ProductList from './ProductList'
-import ProductForm from './ProductForm'
+import CreateProduct from './CreateProduct'
 
 const Product: React.SFC = () => {
   let match = useRouteMatch()
@@ -13,7 +13,7 @@ const Product: React.SFC = () => {
   return (
     <Switch>
       <Route path={`${match.path}/new`}>
-        <ProductForm/>
+        <CreateProduct/>
       </Route>
       <Route path={match.path}>
         <ProductList/>
